getcontainer/requirements.txt
boto3==1.12.10
botocore==1.15.10
colorama==0.4.1
dnspython==1.16.0
docutils==0.15.2
jmespath==0.9.4
pyasn1==0.4.7
python-dateutil==2.8.0
PyYAML==5.1.2
rsa==3.4.2
s3transfer==0.3.3
six==1.13.0
urllib3==1.25.7
aws-xray-sdk==2.4.3
